SectionTableを抜き出す
